[mojo-lang] Mark decls as fully resolved when parsing bodies (#16055)
The existing issue #6701 proposes that the parsing of expressions be
deferred in all contexts to allow the declarations in bodies of others
to be processed first before parsing any expressions. This goes in the
other direction, parsing expressions as they appear but allow lookup in
the body of a decl as it is being parsed.

This fixes #5975 in the sense that FuncOp no longer gets special
treatment: all bodies are treated like the bodies of FuncOp. However,
the hack needed to allow signatures to access input parameters is still
there. This will be fixed in the next PR.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 05350f70b2db63d67ac99d347d13ab90920a50e4