[KGEN] Remove unused/invalid KGEN deps list (#16068)
As part of the KGEN version hash, it uses a fixed declared set of
directories. Some of these no longer exist. Specifically, `HLCFDialect`
is no longer in `Support`: it's in `KGEN/lib` which is already covered
as a top-level dependency for kgen in this list. `HLCFToLLVM` is no
longer in the codebase, so it's definitely not needed in the kgen deps
list.

MODULAR_ORIG_COMMIT_REV_ID: c86086794b95f10360a86a9faec480f1961e2b0b