[mojo-lang] Parse signatures using a signature scope (#16056)
The signatures of structs and functions contain types that can refer to
input or result parameters. Those declarations were added to the ASTDecl
of the struct or function, requiring it to be marked as fully resolved
during signature solution to allow lookup to happen. This was a bit of a
hack. The signatures are actually self-contained name scopes, so this PR
introduces a signature scope when parsing struct and function signatures
and then just forwards the declarations to the ASTDecl corresponding to
the struct or function.

This is a step towards fixing #5975

MODULAR_ORIG_COMMIT_REV_ID: 0479e9bf580b3169ca9825f40bb3fe95dd1024a4