[mojo-lang] Immediately resolve var/let decls in functions (#16058)
The problem with treating local variables in functions the same as other
declarations -- that is, non-lexical resolution -- is that they are not
non-lexical entities: they are SSA values with dominance. Before, this
meant that various parts of the parser required hacky dominance checks
to
generate correct IR, prevent crashes, and emit good errors. Now, var/let
decls inside functions are immediately resolved, removing the need to do
all of this. This does not implement the scoped hash table thing that
would be needed to get rid of the nested ASTDecl hack for variable
scopes.

Fixes #5975

MODULAR_ORIG_COMMIT_REV_ID: 92260cea64ce86f90b35ff04a9cc34d10814463d