[KGEN] Include compiler-rt-cas CAS ID in error message (#16075)
When the `compiler-rt-cas` artifact is not found in the CAS in the
`ExecutionEngine`, add the CAS ID it tried to look up for debugging
purposes as we root cause this problem. While this is an implementation
detail, it's super useful for debugging right now.

This CAS ID is the value from the `cache_files` function that gets
output to the specified `HEADER_NAME` as used in
`KGEN/lib/ExecutionEngine/CMakeLists.txt`. Essentially, we want to
ensure the CAS ID that got put in matches the one that the
`ExecutionEngine` tries to retrieve at runtime.

MODULAR_ORIG_COMMIT_REV_ID: 6db78a419c131084784445a5d4a07beaaf5e9085