[mojo-lang] Make ASTDecl child iteration deterministic (#16013)
This PR switches AST::declsInScope to use MapVector instead
of DenseMap to ensure that we can iterate over the decls in
a deterministic way.

This change also revealed an iterator invalidation bug where we
held dead references to the child map after an insertion (we've
gotten lucky so far to not hit this in practice).

MODULAR_ORIG_COMMIT_REV_ID: a4dea1ed5d72bde3e16a416abcd8716180553edf