[KGEN] Disable name shadowing in `ParameterUseDefGraph` (#16046)
This patch changes the `ParameterUseDefGraph` so that it no longer
allows nested parameter declarations with the same name, i.e. name
shadowing in KGEN IR. The parametric inliner relied on this feature
heavily, and the patch modifies it so that it now ensures unique
parameter names in nested scopes of the inlined function.

MODULAR_ORIG_COMMIT_REV_ID: 24d2daadefaf62176addaab96a6d7883c6c86af2