[Cache] Always insert up-to-date CAS IDs in `cache-mgr` (#16132)
When `cache-mgr` is run to insert the CAS IDs for artifacts, it may not
always run with the latest up-to-date `KGENVersion` string, which is
used to inform the directory to put the CAS contents in, such as

`.derived/.kgen_cache/orc/0.0.0-42904bf96d5b7c1bb524e91b3e7643362f6faba8596e85718f0ef6a8be71fc10-release`

This means that during execution time of `kgen` in the
`ExecutionEngine`, when it looks up the CAS ID, it will not be able to
find CAS artifacts such as `compilerRT`. An example of how to see this
problem:

1. Run CMake and build `ExecutionEngine` target which would tell the
`cache-mgr` to insert the CAS IDs in a particular directory structure
based on the `KGENVersion`.
2. Change something (such as `BlobCache.cpp` contents) which would
affect the `KGENVersion`.
3. Run `check-kgen` (or anything that would invoke the code path through
the `ExeExecutionEngine` and look up the libs such as `compilerRT` from
the CAS). This would result in a failure to find the right CAS ID. The
root cause is because the CAS ID and KGENVersion have changed, but the
`cache-mgr` was not rerun to put the new CAS ID in the right directory
structure.

Fix this problem by having the `cache-mgr` custom targets created by
callers of `cache_files` to properly depend on `KGENVersion`. In turn,
this means when tools like `kgen` and friends depend on
`compiler-rt-cas` target, it propagates that dependency of `KGENVersion`
which would mean that `cache-mgr` is rerun when the KGEN version
changes, which is absolutely needed.

This means one can freely change something that would affect the
`KGENVersion` and just run things like `check-kgen`, and the new CAS IDs
will be inserted in the right filesystem-backed CAS we use, for example.
Specifically, in the `check-kgen` case, it would take an MLIR file in
its unit tests, run `kgen` which would invoke the `ExecutionEngine`, and
find the right CAS IDs when looking up `compilerRT` and friends now. No
more workarounds, no more hoop jumping. It just works - as intended.

MODULAR_ORIG_COMMIT_REV_ID: 03cc1bdec573744fbf46ec1ad7fdc1c82a1684d8