[mojo] Remove unused `mojo -emit` option (#16138)
The existing `mojo` driver has an option, `-emit`, that when specified
would emit an archive file (instead of executing the Mojo program, which
is the behavior when no `-emit*` options are passed on the command
line).

This option is no longer used by any automated tests, and its
functionality exists in the `kgen` tool anyway, so there's no need to
keep it in `mojo`. Remove it to simplify the replacement of `mojo` with
`mojo-driver`.

This work is part of #15442.

MODULAR_ORIG_COMMIT_REV_ID: b1ffa6113a3b3ec6f78a0180b7bbaa476d33e09b