[KGEN] Simplify the API to attach a detailed trace to a chain (#16106)
This both simplifies and optimizes the code to attach a detail
information to a trace. It now elides constructing a trace with a
description if it is not being provided. It also reduces the number of
exported functions in compilerrt by 1.

MODULAR_ORIG_COMMIT_REV_ID: 5e710364cb11711d607f526540562d252452eacc