[mojo-driver] Add `mojo-driver package` command. (#16214)
This patch adds the `mojo-driver package` command. Currently, this only works on single MLIR files where we manually create a `lit.package` because the frontend work is not yet finished. In the future, this will take a directory and generate a precompiled package from that.

The current implementation is that the top-level package is essentially renamed (see the test in test/mojo-driver/mojo-package.mlir for context) - so we discard the top level package name in favor of the one specified on the command line.

MODULAR_ORIG_COMMIT_REV_ID: 828c9fc70e00bf330c051c89bb09d82df548b678