[mojo-lang] Add language support for 'is' and 'is not' operators. (#16272)
This implements these operators in terms of an `__is__` and `__isnot__`
method on the receiver object.  This allows PythonObject and Object to
implement the semantics defined by Python (comparing the object
identity pointers), but also allows types to provie custom behaviors
as well (e.g. object will have an inlined variant representation).

This doesn't provide the library support for this.  This subsumes the
test in PR #15579 (Thanks Alex!) and fixes PR #15578.  This doesn't
declare success for is/isnot on the public tracker, but I'll file an
issue for library support.

MODULAR_ORIG_COMMIT_REV_ID: 8dd4809e9ef17903ac2b38de4f57d6ff410eaece