[mojo-lang] Clean up comma parsing logic. (#16293)
This removes parseListUntil (it is unused) and changes
parseCommaSeparatedList to take a statement indent (like everything
else) to decide when to stop.  This will be needed when we support
things like: `x = 1, 2,` and want the elements to be considered part
of the current statement when they are indented more than x even if
they are on a new line.

MODULAR_ORIG_COMMIT_REV_ID: 71d6daf0938e78a28bcd63fa0859ac93785ad915