[mojo-lang] Fix assignment statement parsing logic, fix tuples. (#16294)
This fixes the handling of `=` and `+=` parsing to use the proper
grammar instead of parsing the left and right sides as expression
akin to a normal binary operator.  This fixes a bunch of issues,
and notably allows tuples to work without parens, a long-standing
friction issue that has been on the roadmap for some time.

This fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: a524d02336a77b8d035dac8e6f54e0d249b45f09