[KGEN] Add `precompiledBodyRef` attr. (#16309)
This patch adds a new attribute to kgen.func - an optional SymbolRefAttr. If this attribute is present, it indicates that the function was precompiled, and that its code is contained in the link directive the attribute points at.

This becomes useful in subsequent patches - we attach this attribute to post-elaboration functions in precompiled binary packages.

MODULAR_ORIG_COMMIT_REV_ID: bbbd31a17f49deaaf2ebfd11d20ba28dee0d6419