[KGEN] Update inliner op kind assumptions. (#16311)
The inliner assumed that when a symbol was looked-up, it was of type FuncOpT, but this is not always the case since before elaboration when we have binary packages, we intermix kgen.generator and kgen.func. This patch updates the inliner to only add an edge to the graph if the op found is of the expected type, if it is not, then we ignore it.

MODULAR_ORIG_COMMIT_REV_ID: 080b2a9db06ce939865bf59b96a5ed2d90659b95