[LLCL] Remove typename array tests (#16361)
The `typeNameFor<T>` unit tests where `T` is an array type is sensitive
to spacing since is depends on `PRETTY_FUNCTION` output which varies
across compilers. Even on Clang, it seems there are some differences on
whether there is a space or not.

It's not super high priority to have these array tests if it's going to
cause a bit of developer headaches. On Clang and GCC, there isn't any
special casing for array types when parsing the output, so it's not a
big deal to skip out on these tests.

MODULAR_ORIG_COMMIT_REV_ID: f029f7653ba3620031a73eeabddbc7fdb83984e6