[mojo-lang] Add warning for transfer from RValue. (#16364)
The transfer operator is intended to end the lifetime of an LValue
or BValue converting it into an RValue.  It is a pointless noop to
transfer from an already-owned value.  As such, add a warning about
this to help nudge users in the right direction.

While here, rename the internal nomenclature from being the "consume"
operator to the "transfer" operator since that's what we settled on
in the public docs.

MODULAR_ORIG_COMMIT_REV_ID: e3091f9469402161e9110244aa07f8a82e6de737