[mojo-lang] Add parser support for declaring global variables (#16346)
This teaches the parser and all the ASTDecl-related logic about
GlobalVarDeclOp and the statement parser about declarations in the
top-level scope. The PR introduces a `GlobalDLValue` that can be coerced
to an `SLValue` or `MBValue` (for `let` globals). This PR moves the
contextual `expr` down to the derived classes of `BaseDLValue`, since
there are contexts where an `expr` cannot be provided for
`GlobalDLValue`.

This allows variables to be declared with destructors and initializers,
but no ability to reference them yet.

MODULAR_ORIG_COMMIT_REV_ID: fd4c8eef493fabda45cbd5e4d45da24f22f22195