[mojo-driver] Move driver basics to Support. (#16357)
This patch moves some of the driver basics to Support so that we can use them for other drivers at modular - specifically, the `modular` installer tool. We want to keep a consistent experience, so it makes sense to use the same infra.

This patch doesn't move mojo-tblgen to support because that's a larger endeavor, so #16356 tracks doing this.

MODULAR_ORIG_COMMIT_REV_ID: 60f8f61bc67cba4328276bc1a02e30bb6ee61469