[LIT] Add handling for precompiled funcs. (Re-apply #16313) (#16368)
This patch adds the handling for lit.func that have been externalized. During the LowerLIT pass, we replace the lit.func with its fully-elaborated bytecode. Because the func may be exported, and therefore have a different name in its compiled form, we also have to pipe a map down the stack to be used when renaming calls.

MODULAR_ORIG_COMMIT_REV_ID: 25ec9e678dfe106007c5fe2409265a2bcce10064