[KGEN] Externalize precompiled functions. (#16314)
This patch adds the capability to fully externalize precompiled functions. If a function has already been compiled by a binary package, we should not re-compile the code for it! This patch adds a pass that runs after inlining that removes functions and replaces their uses with pop.external_call that correctly reference the link directive that contains the archive for the precompiled binary package in question.

MODULAR_ORIG_COMMIT_REV_ID: 7c34ea475e46978714f833c8165314df1491e88e