[KGEN] Allow DCE to remove unused global variables (#16363)
The notion of global variables are that they are removed from the
program if they are unused, meaning that their initializers cannot be
relied on to have certain side effects. This is a temporary change in
order to align the notion of exports with the current model (plus
`sliceDependencies` already removes unused globals).

MODULAR_ORIG_COMMIT_REV_ID: 703c61fad1ffe6ea0a35c72b7cb8a5ee5ce700bd