[mojo] Remove codegen trigger (#16456)
Remove what the comments call a "sketchy" method of ensuring that
codegen occurs even when no symbols are exported. The comment claims
that tests need to be udpated for this to work, but tests actually pass
with this code removed, as far as I can see.

`mojo` contains several checks for exported symbols (one that the module
contains symbol interface ops, and one that `main` is exported), so it
doesn't need this additional check. `kgen` doesn't seem to explicitly
check for this case, so replace the workaround with an error and add a
test case.

Closes #10893.
MODULAR_ORIG_COMMIT_REV_ID: b7ab1fe4e3477684cefd065de61db59154c1722d