[LSP] Add initial version of a symbol index (#15985)
- Add a listener interface that is used in the parser to notify the LSP
when var/let decls are resolved, as well as when references to these
decls are also resolved.
- Implement this interface in the LSP to populate a new symbol index
with variable decls and references.
  - This index will be used to implement hover and goToDefinition.
- This index is not handling cross-file references. This will be fixed
later.
- This index is not storing the symbols of non-opened files. This will
also be fixed later.
- Added a few handy functions in MojoASTDeclRef and fixed
MojoASTTypeRef, which seems not to have worked in the first place, but
at least it served as good scaffolding. Feedback is welcome.

MODULAR_ORIG_COMMIT_REV_ID: 7544d7832a5bd020f96cdd83e3c8071951c59540