[KGEN] Invoked JIT'd functions through the ORC RT (#16481)
One problem with this PR is that it crashes when calling `def main() ->
None` because `orc_rt_run_program` is hard-coded to expect `int
main(int,
char**)` as the signature. It handles `fn main()` fine however.

This also does not fix the initializer issue. Although `dlinitialize` is
now getting called through the ORC RT, the ExecutionEngine is stripping
the init sections of the object files because the binaries of the
symbols are being ripped out of the standalone archive and fed into the
JIT, meaining the resulting JITDylib has all the necessary objects but
is still missing the init sections.

MODULAR_ORIG_COMMIT_REV_ID: 7b642f67d4362c462284c4ac4f333f9e98dbabca