[Support] Remove the SIMD.h file since it's no longer used (#16607)
The SIMD.h file was introduced to allow folks to write kernels that use
vectorization in a portable way in C++ without having to resort to
intrinsics. While that was a good idea 8-10 months ago since we did not
have a language it is no longer valid today, since we have Mojo. The use
case for this file has thus disapeared and so we can just remove it.

MODULAR_ORIG_COMMIT_REV_ID: 0eed614819a88890a688f8debb37fb1910f93a91