[mojo-lang] Introduce PrettyCrashReporter, NFC. (#16626)
This enhances the Mojo parser to push a PrettyCrashReporter when
parsing at statement or declaration.  If a crash happens in that
scope, it prints out the source range of tokens that was consumed
while at it.  For example, on the bug in Issue #16606 it prints
out:

```
0.	Program arguments: /Users/clattner/Projects/modular/.derived/build-debug/bin/kgen -gen-lib /Users/clattner/Projects/modular/Kernels/mojo/Mogg/MOGG.mojo -o /Users/clattner/Projects/modular/.derived/build-debug/Kernels/mojo/Mogg/MOGG.lib.mlirbc -O3 -debug-level=none
1.	Crash resolving decl body at loc("/Users/clattner/Projects/modular/Kernels/mojo/LinAlg/Matmul.mojo":1193:5)
    >>     fn run(
           ^......
    >>         c: NDBuffer[2, config.shape_c, c_type],
    >>         a: NDBuffer[2, config.shape_a, a_type],
    >>         b: NDBuffer[2, config.shape_b, b_type],
    >>         elementwise_epilogue_fn: fn (GemmShape, GemmShape) capturing -> None,
2.	Crash parsing statement at loc("/Users/clattner/Projects/modular/Kernels/mojo/LinAlg/Matmul.mojo":1243:9)
    >>         matmul._outer_k_loop()
               ^.....................<
```

This will make user-reported bugs incrementally easier to deal
with, and probably can tie into telemetry.

MODULAR_ORIG_COMMIT_REV_ID: 8f785984dbf645cb320a3578e31a06bc424047bd