[mojo] Restructure post-elaboration bytecode ref to be the entire module (#16620)
Right now we just store the individual function body, but using the
full module allows for pulling in dependencies (to better enable
interpretation/inlining/etc.). With this, the inflation of
post-elaboration functions is shifted to a specific pass that can
run at the end of the LIT lowering pipeline (which also ensures that
types and things in the input module are lowered before we start
pulling in bytecode that expects things to be that way).

This also adds some additional verification missing for the post
elaboration/external function functionality.

MODULAR_ORIG_COMMIT_REV_ID: 8317ac0de6c682cc31b0091c54fe2f186e6fa937