[LSP] Add a --suspended flag for debugging the lsp (#16685)
This is useful for debugging the lsp, because it initializes very
quickly. I also plan to write a simple client tool that can be used for
tests and scripting, but that will happen later.

- This new flag causes the lsp server to stop right after the transport
and the logger have been set up. The client (vscode) waits undefinitely
until LLDB attaches to the server and resumes the execution.
- I also added a VSCode command that restarts the server in a suspended
state, which makes it easier to debug crashes.

MODULAR_ORIG_COMMIT_REV_ID: a0b722b5ed6fdbeaa82009e760338637c02544d2