[Mojo][Docs] Upgrade the programming manual and HelloMojo notebooks (#16076)
This is the final change to resolve
[Internal link] so that the Mojo
Programming Manual is completely written in a Jupyter notebook and most
(not all) code is runnable. In doing so, the "Hello Mojo" notebook is
now a much shorter, all new notebook (the original HelloMojo was
essentially a copy of the programming manual, but was not up to date
with changes in the manual). So now, in addition to a shorter "Hello
Mojo" notebook, the Mojo Playground also has a word-for-word runnable
version of the Mojo Programming Manual.

The bulk of this work was actually performed with previous PRs. This PR
simply moves the `programming-manual.ipynb` file and new
`HelloMojo.ipynb` file to the `public` directories:
- Switch the HelloMojo notebook to a simplified language intro.
   - Original PR: [Internal link]
- Add notebook version of the complete programming manual. That is,
convert the .md manual to an .ipynb notebook.
   - Original PR: [Internal link]
- And making these switches requires some special-treatment in the docs
CMake so we move the notebook-based programming manual to the same URL
location as the original manual (instead of leaving it in the
"notebooks" directory).

If you want to review the docs as rendered (recommended):

```
git checkout -b mojomanual2 origin/scott/mojomanual2

build docs-website-mojo

quarto preview .derived/docs-website
```

When quarto is done rendering, it will open the rendered site in your
browser.

Resolves:
[Internal link]
[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 5b4b7333e881dddaeed40b97c5b8a176b1f03cc3