[mojo] Extend scope of diagnostic handler (#16733)
`mlir::SourceMgrDiagnosticHandler` is a `ScopedDiagnosticHandler`, and
needs to remain in scope as long as it ought to handle diagnostics from
MLIR. When splitting up the `run` function in code review for #16478, I
limited its scope to the `compileModule` function, and so diagnostics
emitted during the `executeModule` function's execution did not appear
correctly.

Instantiate the handler within the main `run` function in order for its
scope to last the entire duration of `mojo run`.

MODULAR_ORIG_COMMIT_REV_ID: 0aee50beec57e915d8c836728d20827d15225ca6