[KGEN] Remove alias from ExportOp (#16736)
This PR removes the alias field from ExportOp. It uses the
soon-to-be-renamed "postElaborationFuncName" as the "export alias" of
the symbol and performs the rename early. This means that *LowerLIT is
producing the symbol names in the eventual binary*, which removes a
bunch of extra complexity of juggling around an alias.

This also required a minor change to the REPL implementation to generate
a unique top-level expr func name for each expression, rather than
always using the same one.

MODULAR_ORIG_COMMIT_REV_ID: 9113ec833c89669b470e056426ca19f8d1978c99