[KGEN] Fix potential deadlock in elaborator (#16778)
I don't believe this has been observed in practice because usage of
recursion in our Mojo codebase is extremely limited and only comes in
a handful of shapes on most of which the deadlock is impossible.
TSAN warns about it, and it's because the locks on `bindings` are
acquired in different orders in the code. This PR ensures they are
always acquired in the same order: parent to callee, which is reversed
when breaking recursion (callee can become the parent).

MODULAR_ORIG_COMMIT_REV_ID: a63dd4a67961f6911cf833c6eeb69cf5e347d5c1