[Support][LLCL] Introduce `-static` versions of the Globals libraries. (#16803)
This patch introduces MSupportGlobals-static and LLCLRuntimeGlobals-static. These libraries are useful for things like the `modular` tool, which won't (for the forseeable future) interact with other things that might load these globals and need to pass objects across a `dlopen` boundary.

The suffix `-static` was chosen to (a) keep the original as the default, and (b) because it's slightly harder to type, so it forces the user to check themselves.

MODULAR_ORIG_COMMIT_REV_ID: 6cdc990b956e630e2dda886b4e24c97c008abcb0