[mojo-lang] Don't loop forever in `getAutoImportPaths`. (#16814)
When we don't have MODULAR_PATH or MODULAR_DERIVED_PATH, we would loop forever because `/` is not 'empty' per se, and `/` has a parent, it's just itself. This patch changes the check to see when the parent path would be the same as the current path.

MODULAR_ORIG_COMMIT_REV_ID: 6faca4b90ca4927e13a21b0643fc8bdad7398fdb