[KGEN] Patch another TSAN complaint (#16831)
TSAN warns that accessing `inode->error` in `completeImplNodeProcessing`
is a data race. In reality, it's a benign race. The code will be correct
regardless of whether the access reads true or false. There is a way to
tell TSAN that a race is benign, but that's too much effort to set up at
the moment.

Just use an extra atomic flag to synchronize this part.

MODULAR_ORIG_COMMIT_REV_ID: 11a96b360588490e12495d0b6437fc85decc9884