[LLCL][Support] Define MODULAR_BUILDING_LIBRARY (#16847)
In #16834 we discovered that you cannot apply __declspec(dllimport) to a function definition. The functions that errored all had MODULAR_CXX_EXPORT tagged on them, which resolves to __declspec(dllimport) if MODULAR_BUILDING_LIBRARY is not defined, and __declspec(dllexport) if it is (this behavior feels...questionable, but let's roll with it for now). We set MODULAR_BUILDING_LIBRARY because according to the MSVC docs, you *are* allowed to set __declspec(dllexport) on a function definition.

MODULAR_ORIG_COMMIT_REV_ID: e873237296a4a07d863a520ca55f13b67d9581cd