[KGEN] Allow elaborator to concretize locations (#16601)
This patch adds a hacky solution that allows the elaborator to concretize locations. The implementation relies on traversing and concretizing all op and block argument locations after an elaborator work item is complete, since `ParameterUseDefGraph` is currently not capable of identifying parameter uses in locations. In particular, regions of `lit.try` (and potentially other ops) need to be visited manually. To mitigate performance effects, this is turned on only if compiling with debuginfo.

MODULAR_ORIG_COMMIT_REV_ID: ea2320a5503234ade9dd9281bc924fcf8ec2535e