[mojo-lang] Make `@export` an early decorator (#16876)
This decorator was the first "late" decorator added (a bit of a hack),
because it required information about the function before it could be
applied, since it had to generate a `kgen.export` operation. Now that
export is a property of the FuncOp, it can run as an early decorator.

MODULAR_ORIG_COMMIT_REV_ID: 187e41cfb3fd0ae2dfa6cbb0cdf81edd470a022d