[Cache] Use `readFileAtomically` in `FilesystemBackend`. (#16910)
This patch moves the FilesystemBackend to read the file atomically using the new readFileAtomically API. This required a little bit of finagling in that we have to pull a BufferRef *out* of the lambda, and we couldn't return it *from* the lambda because it's not itself in Support. This can/should be revisited if we decide to move Cache::BufferRef to Support, but for now this should work fine.

MODULAR_ORIG_COMMIT_REV_ID: e84eb35cf45f2ae661b93ddef980a1c53944bfb9