[Cache] Don't re-insert things into the FilesystemBackend. (#16911)
For the FilesystemBackend, insertion is pretty expensive since we have to get a lock file and write the contents. Take advantage of the fact that this is a CAS, and notice that if we already have something in the CAS, we don't need to write it again.

MODULAR_ORIG_COMMIT_REV_ID: f9696c6b5fbd1cd697910dbe6bb6620a62892a9a