[Support] Add 'canonical' JSON writing. (#16961)
This patch adds canonical JSON writing, which we will need for the
installer. It follows RFC8785 as much as possible - the only deviation
is that we use UTF8 because (a) everything we care to express is
expressible in UTF8, and (b) it's *much* easier. We could fix this in
the future if we wanted to, but for now I believe it's simpler to just
enforce UTF8 everywhere.

MODULAR_ORIG_COMMIT_REV_ID: 23d083e4e58540eb87d0a23138f953b4a37c5769