[mojo-driver] Add a new mojo `build` command (#16833)
This command allows for generating a standalone executable
from a given mojo file. This command should eventually be
merged with the existing `package` command when we have a
way to define mojo projects. An executable is built by generating
a static archive and linking it with KGENCompilerRT-static, currently
using the platform c++ compiler (to simplify pulling in the necessary
goop wanted by KGENCompilerRT, we'll eventually want to shift this
to calling the linker directly).

MODULAR_ORIG_COMMIT_REV_ID: 605357eb306fd827444dc416307d9772796e2d21