[KGEN] Ensure `target-cpu` and `target-features` propagate to LLVM (#17052)
Clang attaches them to function attributes of every function in the
LLVMIR. We don't do this, but because it's passed through TargetMachine
to the JIT, I think it was fine. In any case, do this to match Clang's
behaviour.

MODULAR_ORIG_COMMIT_REV_ID: 858243910864cfca37f976ee928f4475a187ccb1