[mojo-tooling] Build a binary mblack as part of `mojo` (#17078)
This better enables the use of `mojo` as a standalone tool,
and is a necessary pre-requisite for switching some of our
developer tooling to use `mojo` (e.g. vscode) as it removes
dependencies on special python environments/etc. This
is also a necessary step for packaging `mojo` for distribution.

MODULAR_ORIG_COMMIT_REV_ID: 597045b3744ba6144da192e54a937dc6280a3ff0