[cmake] Fix `mblack` build. (#17144)
For building `mblack` binary we assume that the environment in which
cmake is configured is where `pyinstaller` is installed. This may not
always be true. Introduce a MODULAR_PYTHON_EXECUTABLE variable for these
scenarios.

In windows `;` was being as used as separator within pyinstaller
command. This is also the string separator for cmake. So cmake will
split the command where we don't want it to split. To account for this
escape the semi colon.

While there use the new macro from AddPython module for python
discovery.

MODULAR_ORIG_COMMIT_REV_ID: 6b7eb1aa4fa70c58b781a4ba14c582de06a4af06