[Telemetry] Fix OTel histogram Record call (needs context) (#17161)
OTel's Histogram::Record() needs a Context object as the second argument
(which I was not passing). Now that we have code that actually creates
histograms this error came up. Adding telemetry unit tests is on my TODO
list.

NOTE: Counters don't need Context, and it's not clear to me yet why the
context is needed for histograms. Currently I'm just passing an empty
context to fix the build issues, but I have to look into whether this is
relevant to us or not. I think the context is for linking metrics with
distributed traces, which we don't care about now.

MODULAR_ORIG_COMMIT_REV_ID: 895c88d2de13081ee7850708be312561e457e606