[Packaging] Skip binary mblack build for packaging (#17182)
This skips the binary `mblack` build for packaging builds.
This is needed to package Python wheels with the Python distributed with
manylinux.
That is because by design the manylinux Python is statically linked.
However, PyInstaller is not compatible with a statically linked Python,
and rather requires libpython.so to be present.

So to unblock wheel packaging this change skips the binary `mblack`
build for packaging, since that is not a dependency of the
`modular.engine` Python package.

Testing:
- This was manually tested to produce a working `modular.engine` Python
  wheel in a manylinux Docker image.

Related PRs:
- #17078 introduced the binary `mblack` build as a step towards
  packaging `mojo` for idstribution.

Alternatives and future work:
- Another approach here is to have manylinux build a pyenv that is
  dynamically linked.
  Then pass that dynamically linked pyenv Python to PyInstaller here.
  This will be addressed in a follow up tracked by #17186.

MODULAR_ORIG_COMMIT_REV_ID: 01bf4d98ed271a60f76b34ede49bce5b2850ca38