[API] Realign the DType names of the API functions to Mojo (#14887)
In [Internal link] we made the names of
the dtype fields align with Mojo. We changed the names of the fields for
the mojo dtypes to more closely align to numpy. This performs a
realginment (essentially a revert of #12160).

MAX_PYTHON_ORIG_REV_ID: c1e9d0a9ec74e2257c1f5e2bc548d0605c7cd5d5