[mojo-lang] Change how builtin types get resolved to elim crashes. (#17213)
The previous approach had `getBuiltinXXXType()` return null when
there is an error resolving it.  This led to bugs in weird cases
that led to a crash which wasn't very friendly.  Fix this by
returning `TypeCheckError` type instead, which is the signifier for
"we already diagnosed a type checking bug".

This also introduces an `ASTType::isTypeCheckErrorType()` method
which cleans up some tests for this type.

MODULAR_ORIG_COMMIT_REV_ID: d8737868cc986a5c5e7b61b7203dd19d059671b5