[Docs][Mojo][LSP] Extract out the DocString's parsed decls into reusable components - 1 (#17120)
DocString.cpp parses several decls and produces a very readable json
object that can be used for generating documentation, and it turns out
that the LSP needs to do basically the same for hover dialogs. And not
only that, LLDB will need to do the same as well when pretty printing.
Given this, I'm extracting these decls, starting with FuncOp, and
creating a set of structs that I'm calling ASTDeclViews. They could be
used by API consumers that don't want to go through the hassle of
undersanding how to extract every piece of information from Ops and
Attrs. Besides that, these view objects are meant to be serializable to
several formats, e.g. can be converted to JSON, can produce code
snippets of their declarations, and can generate markdown docstrings.
Finally, centralizing this parsing logic in one place will be beneficial
to many language tools.

- Define a main FunctionDeclView, along with other child objects, like
ArgumentDeclView.
- These are WIP structs and will be improved. For example, it uses
std::string for its documentation entries and for types. I want to keep
this PR minimal, so I'm not changing that.
- Move the logic in DocString.cpp that produces the function json blob
to a new ASTDeclView.cpp file. This time it generates the struct, which
has a method to be converted to json.
- Reuse FunctionDeclView in DocString.cpp and in MojoServer.cpp to print
code snippets and docstrings.
- The doc tests pass after simply changing the order of json entries.
The reason is that the json serialization is now dumping the keys in
alphabetical order.

MODULAR_ORIG_COMMIT_REV_ID: ea4cbae507b88c37d55235026878a189b1293e62