[Telemetry] Add custom MetricReader (#17252)
OpenTelemetry's PeriodicExportingMetricReader is not suitable for our
use cases. It uses a background thread that periodically checks if there
are metrics to export. The problem is that when our Runtime shuts down,
I don't see a way to wait for this reader and its thread to export
metrics. So, it can easily happen that not all metrics are exported at
Modular process exit, and also it's common to see "MetricReader::Collect
invoked while Shutdown in progress!" messages dumped to stdout.

Maybe the proper solution is to extend PeriodicExportingMetricReader,
but for now it's simpler to define our own reader. The only caveat is
that we need to explicitly call it to collect and export metrics.

See comment in ManualExportingMetricReader for more details.

MODULAR_ORIG_COMMIT_REV_ID: e54efc3bcfc6ba0f7c4829c7aa945186ae69bc7a