[mojo-jupyter] Fix display_start/display_end variables in matplotlib (#17313)
The values for these variables was mistakenly not updated when
shifting over to a config file. Will followup with a test that checks
we actually emit these (which is something the original PR was
intended to unlock) so this doesn't happen again and we can
unbreak the playground.

MODULAR_ORIG_COMMIT_REV_ID: 37e5be49ef9439982e23090a3c7d420d40f5b57a