[MOGG] Adds a jit-able where (#17303)
Adds an implementation of where and the shape function which we can jit.
This is essentially the same as the shim implementation sans the
dispatch. This just adds the ability and tests which target it. It does
not yet target it by default or mean we can remove the shim

NN_ORIG_REV_ID: 3be6c7ce1f2f1569af209de3972d9a18a01ba947