[BlobCache] Robustify FindItemThatExistsThenClear test (#17365)
This particular test occasionally fails in CI. To try to tickle the
failure locally I extended the test to support both repeated runs and
concurrent insertions/finds. However while doing that realized the
flakiness is most likely just that the test implicitly assumes
andThenSync awaiters for an AsyncValue will be run when awaiting on that
chain returns. So robustify the cest for that too.

Let's see if it shows up again as a flake in CI with this.

MODULAR_ORIG_COMMIT_REV_ID: dabd9f7eac4da49a18c8034b338602886ff4b9a4