[mojo-driver] Add initial support for "generic" packages (#17295)
This allows for compiling a package on-demand when imported
into a module with a target incompatible with the one used to
build the package. This is done by attaching the pre-elaboration IR
when building the package, which gets selected during lowering
depending on the target used when importing the package. This
gives more flexibility for package authoring, because it allows for
specializing for a specific target but still being able to use the
package on other targets.

MODULAR_ORIG_COMMIT_REV_ID: a09e576b1444a242e83e81e43fb4179f4c9075d3