[mojo-lang] Move createLowerPreElaboratedLIT to the elaboration pipeline (#17381)
This pass produces target specific IR, based on the current target, which breaks
the contract of the pre-elaboration pipeline. Move it to the elaboration as that
is the start of where the module is specialized for the current target.

MODULAR_ORIG_COMMIT_REV_ID: aac04bf07168c71471ef7037b1f8d3a759855362