[Build] Fix non-native builds for arm64 systems (#17429)
This works around some weaknesses in the mojo march processing (see
#17426) by essentially using a different flag for clang and mojo. The
arm64 deliverable for the EAP is weird though since we only want to
support neoverse-n1 which is the lowest common denominator for the
graviton systems and so march=native is fine. But this PR goes a step
further and uses any armv8.2-a as the common denominator arm64 system.

There might be performance impact because of this. See page 11 in
https://www.stonybrook.edu/commcms/ookami/_pdf/Linford_OokamiUGM_2022.pdf

Closes #17392
MODULAR_ORIG_COMMIT_REV_ID: 0a090feeb65dd3c8001b56913503855671ffd2a9