[MGP] Allow empty feature lists and wildcards for triple & cpuArch (#17454)
On MacOS the CPU features list is empty but my string hacking was not
tolerant of that. (Again, it is a shame TargetInfoAttr does not use an
ArrayAttr<StringAttr> for features so it is 1:1 with HostMachineInfo, a
task for a rainy day).

To allow that to be tested, also support empty triple and cpu strings to
denote a wildcard in the runtime check.

---------

Co-authored-by: Mark Shields <mbs@marks-air.lan>
MODULAR_ORIG_COMMIT_REV_ID: ba9691ae6d4fc1fdf18de67e222673a964eb27c8