[Telemetry] Add file exporter (#17460)
This adds the option of exporting metrics and logs to a file, using the
OStream exporters in opentelemetry-cpp.

Note that this is most likely only suitable for debug and dev purposes
(which is why this is being added). I don't think the output is in a
suitable format for reading back and sending.

In this PR, if `MODULAR_TELEMETRY_URL` environment variable does not
exist, metrics and logs will be dumped to
`$MODULAR_DERIVED_PATH/telemetry.log`. To be done as follow-up is to get
these options and paths from a config file.

MODULAR_ORIG_COMMIT_REV_ID: efab581d0de1d7492b0b39c59351eb2b06597d96