[TF] Super scrappy op coverage stats recording (#16542)
This follows the file output model from
[Internal link] It's just a start -
we can expand/change more in the future.

The logging is controlled by two env vars that users can set:
* `MODULAR_STATS_FILENAME` - a file to append stat reports to (one for
each model load)
* `MODULAR_STATS_ALWAYS_RECORD` - when this is set and the file name is
not, a report will always be output to a temporary file that users can
collect and send to us

Open issues (to be addressed at a later time):
* the report now includes op counts broken down by op name and
signature, including attributes; because this can potentially expose
unwanted data, the control for enabling these statistics is left to the
user, and is off by default
* more consistent means of controlling the output file (although this
will be moot once the telemetry reporting is available)
* a model name is *not* include in the report; turns out the file name
(the best indicator) is a bit more involved through route through the
CXX driver API

Fixes #16494.

cc @modularml/framework-frontends FYI - this is just a temporary
solution, we want to decide whether this remains a utility to be used
across lowerings (this implementation) or as a separate pass

MODULAR_ORIG_COMMIT_REV_ID: 7ccff716dbea490d0d32df673fae32fb3c951e68